+++ patzilla-ui/navigator/components/permalink/permalink.js
@@ -128,6 +128,8 @@ PermalinkPlugin = Marionette.Controller.extend({
             baseurl = baseurl.replace('patentsearch', 'patentview');
         } else if (_.string.contains(host, 'navigator')) {
             baseurl = baseurl.replace('navigator', 'viewer');
+        } else if (_.string.contains(host, 'patbib')) {
+            baseurl = baseurl.replace('patbib', 'patview');
         } else if (!_.string.contains(host, 'localhost')) {
             baseurl = baseurl.replace(host, 'patentview.ip-tools.io');
         }