+++ CHANGES.rst
@@ -10,6 +10,7 @@ Development
 - Fix export with non-ascii characters accidentally slipped into patent numbers
 - Improve styling capabilities
 - Remove defunct upstream database
+- Add ``patview`` to the list of accepted patentview subdomains
 
 
 2019-03-09 0.164.1